+++ src/main/java/io/github/enelrith/bluebay/users/repositories/UserInformationRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 /**
  * Repository for {@link io.github.enelrith.bluebay.users.entities.UserInformation}
  */
-public interface UserInformationRepository extends JpaRepository<UserInformation, Integer> {
+public interface UserInformationRepository extends JpaRepository<UserInformation, Long> {
 }
